+++ H/Nodes/Token.h
@@ -43,7 +43,7 @@ public:
 	//for tokens
 	Token(const Token& t, long long f) { *this = t; Flags = f; }
 	Token(const Token& t, int s) { *this = t; Size = s; }
-	Token(long f) : Flags(f) {}
+	Token(long long f) : Flags(f) {}
 	Token(Node* n);
 	Token(long long f, int s) : Flags(f), Size(s) {}
 	Token(long long f, string n) : Flags(f), Name(n) {}
